requirements.txt
Django==1.9.6
django-webpack-loader==0.3.3
djangorestframework==3.3.3
